Disable environment.noXlibs in the default Hackworth NixOS env.
Reasons:

- This occasionally causes very subtle and difficult-to-debug
breakage. The most recent example is causing one of git's compgen
tests to fail. It took me quite a while to figure out why that was
happening, and yes, it's related to this configuration setting,
somehow.

- I get the impression that upstream doesn't test this setting very
much.

- I did it to save build times for our headless NixOS machines, but
ironically, I think it actually causes longer build times because
packages that are affected by it end up missing the upstream nixpkgs
binary cache.