Remove nix patch that works around stack overflows in some cases.
Unfortunately, it's not compatible with recent versions of Nix. Note
that this might mean that some of our packages are not buildable on
the Hydra, though I have only seen this issue so far with the nixpkgs
machinery for building NixOS ISO images, which we can live without for
the time being.