Remove hydra override.
Something that's really annoying about Nix Flakes is that, sometimes,
if you add a dependency to a flake that is also a flake, then you
often have to deal with the dependency is downstream dependencies.

In this case, flakes that were using the hacknix flake were
complaining that hacknix's Hydra flake had an undefined input, meaning
that we had to define that input even for flakes that don't use Hydra.

Therefore, we simply remove the Hydra flake from hacknix and we'll
have to add our patches to the projects that actually need Hydra. :(