chore(nix): Bump dependencies.
Also, remove the workaround for
https://github.com/LnL7/nix-darwin/issues/373 as our nix-darwin fork
now has a patch for it (see
https://github.com/LnL7/nix-darwin/pull/379).

Note that this nix-darwin rev also removes the `nixUnstable` hacks we
were using previously so that we could use `nixStable` as the system
version, but `nixUnstable` for flakes support in the user's
environment and for `darwin-rebuild`. `nixStable` is now 2.4, so it
should be safe to use it for all scenarios we want to support.