fix: Add a `login` to netrc credentials.
We've been using this Vault Agent template for Cachix, and it has
worked fine until recently, when we started having issues using a
private Cachix cache. The issue appears to be related to

https://github.com/cachix/cachix/issues/406

which is worked around by adding a nonce login name. Furthermore, it's
likely that other netrc clients will need the login, so we now require
it.