ci: Remove Buildkite post-build hook.
Prior to this change, we used a Nix post-build hook to write all
derivations built by this flake into our public Cachix cache. Here we
remove that, as we'd rather configure out builders to push to Cachix
automatically, because:

1. it's much faster than a per-derivation post-build hook; and
2. it obviates the need to give the Buildkite agent write access to
Cachix.

However, at the moment, we only have a single builder, and that
builder is configured to push local Nix store writes to our private
Cachix cache (we can choose only one). Therefore, for the time being,
this flake's derivations will not be cached in our public Cachix
cache. We'll fix this at a later date. (I doubt anyone's using it,
anyway.)