chore: override `vector`
For some reason, `vector` often breaks upstream on `aarch64-darwin`;
e.g.:

https://hydra.nixos.org/build/285322715

However, upstream only seems to consider the `x86_64-linux` build for
gating admission. We have downstreams that depend on `vector` on
`aarch64-darwin`, so now we override it.

This initial commit is `v0.44.0`, which is about to enter upstream
here:

https://github.com/NixOS/nixpkgs/pull/373601

Note that the upstream changes fail on `aarch64-darwin`. However, we
were able to work around this simply by adding `zlib` to `buildInputs`
on macOS. Maybe this is the long-term fix, though it's not clear why
upstream hasn't already done this.

Signed-off-by: Drew Hess <src@drewhess.com>