chore(nix): enable the macOS sandbox, with caveats
It seems like most of the macOS sandbox issues have been resolved,
except for the 64K sandbox program limit.

Here we enable the sandbox, but in order to work around this limit, we
add `/nix/store` to the allowed paths by default. (Presumably the
sandbox compiler is smart enough to replace multiple `/nix/store/...`
paths with a single `/nix/store` parent path.)

This is less than optimal, but much safer than disabling the sandbox
entirely, and probably safer than disabling the sandbox on a
per-package, as-needed basis, as well.

Ref:
https://github.com/NixOS/nix/issues/4119#issuecomment-2561973914

Also see:

https://github.com/NixOS/nixpkgs/pull/346945
https://github.com/NixOS/nix/issues/6836
https://github.com/NixOS/nix/issues/4119
https://github.com/amarshall/home-manager/commit/d7319b754802fe5f9a5c2b7db375a3331779d27d
https://github.com/NixOS/nixpkgs/issues/366245
https://github.com/NixOS/nix/issues/2311

Signed-off-by: Drew Hess <src@drewhess.com>